requirements.txt
numpy>=1.22.0
Pillow>=9.0.0
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.15.0
hdbscan>=0.8.28
umap-learn>=0.5.2
scikit-learn>=1.0.0
matplotlib>=3.5.0
opencv-python>=4.5.0
scikit-image>=0.19.0
sentence-transformers>=2.2.0
requests>=2.25.0 